move gcloud into connections
