+++ dbt/adapters/bigquery/connections/_gcloud.py
